Changed javascript to typescript

+++ mozi-backend/utils/review.ts
@@ -1,9 +1,12 @@
-const tokenModule = require("./token")
+import { MyContext } from "../server";
+import { determineRole, getToken } from "./token";
+import { Review, Token, User } from "./types";
+import { getCurrentUser } from "./user";
 
-function getReviews(_, context) {
+export function getReviews(_:any, context:MyContext):Promise<Review[]> {
   const sql = "SELECT * FROM review";
   return new Promise((resolve, reject) => {
-    context.db.all(sql, [], (err, rows) => {
+    context.db.all(sql, [], (err: any, rows: Review[]) => {
       if (err) {
         reject(err);
       }
@@ -12,10 +15,11 @@ function getReviews(_, context) {
   });
 }
 
-function getReviewById(id, context) {
+export function getReviewById(id:string, context:MyContext):Promise<Review> {
   const sql = `SELECT * FROM review WHERE review.id = ?`;
+  console.log(id)
   return new Promise((resolve, reject) => {
-    context.db.get(sql,[id], (err, rows) => {
+    context.db.get(sql,[id], (err: any, rows: Review) => {
       if (err) {
         reject(err);
       }
@@ -24,10 +28,10 @@ function getReviewById(id, context) {
   });
 }
 
-function getReviewsOfUserForMovie(user_id, movie_id, context) {
+export function getReviewsOfUserForMovie(user_id:string, movie_id:string, context:MyContext):Promise<Review[]> {
   const sql = `SELECT * FROM review WHERE review.user_id = ? AND review.movie_id = ?`;
   return new Promise((resolve, reject) => {
-    context.db.all(sql, [user_id,movie_id], (err, rows) => {
+    context.db.all(sql, [user_id,movie_id], (err: any, rows: Review[]) => {
       if (err) {
         reject(err);
       }
@@ -36,10 +40,10 @@ function getReviewsOfUserForMovie(user_id, movie_id, context) {
   });
 }
 
-function getReviewsOfMovie(movie_id, context) {
+export function getReviewsOfMovie(movie_id:string, context:MyContext):Promise<Review[]> {
   const sql = `SELECT * FROM review WHERE review.movie_id = ?`;
   return new Promise((resolve, reject) => {
-    context.db.all(sql,[movie_id], (err, rows) => {
+    context.db.all(sql,[movie_id], (err:any, rows:Review[]) => {
       if (err) {
         reject(err);
       }
@@ -48,11 +52,11 @@ function getReviewsOfMovie(movie_id, context) {
   });
 }
 
-function createReview(review, context) {
+export function createReview(review:Review, context:MyContext):Promise<Review> {
   const sql = `INSERT INTO review (id,rating,description,movie_id,user_id)
     VALUES (?,?,?,?,?)`;
   return new Promise((resolve, reject) => {
-    context.db.run(sql,[review.id,review.rating,review.description,review.movie_id,review.user_id], (err, rows) => {
+    context.db.run(sql,[review.id,review.rating,review.description,review.movie.id,review.user.id], (err:any) => {
       if (err) {
         reject(err);
       }
@@ -61,22 +65,23 @@ function createReview(review, context) {
   });
 }
 
-async function updateReview(review, context) {
-  const token = await tokenModule.getToken(user, context);
+export async function updateReview(review:Review, context:MyContext):Promise<Review> {
+  const user:User = await getCurrentUser(context)
+  const token:Token = await getToken(user, context);
   if (!token) throw new Error("No Token");
-  if (token.token === context.req.headers["auth-token"]) {
+  if (token.token === context.req.headers['auth-token']) {
     const updatedReview = await getReviewById(review.id, context);
     const returnReview = {
       ...review,
-      movie_id: updatedReview.movie_id,
-      user_id: updatedReview.user_id,
+      movie_id: updatedReview.movie.id,
+      user_id: updatedReview.user.id,
     };
     const sql = `UPDATE review SET rating=?, description=? WHERE review.id = ?`;
     return new Promise((resolve, reject) => {
       context.db.run(
         sql,
         [review.rating, review.description, review.id],
-        (err) => {
+        (err:any) => {
           if (err) {
             reject(err);
           }
@@ -87,18 +92,20 @@ async function updateReview(review, context) {
   }throw new Error("Unauthorized!");
 }
 
-async function deleteReview(id, context) {
-  const token = await tokenModule.getToken(user, context);
+export async function deleteReview(id:string, context:MyContext):Promise<Review> {
+  const user:User = await getCurrentUser(context)
+  const token:Token = await getToken(user, context);
   if (!token) throw new Error("No Token");
-  const role = await tokenModule.determineRole(context);
+  const role = await determineRole(context);
+  if(role === undefined) throw new Error("Role not found")
   if (
-    token.token === context.req.headers["auth-token"] ||
+    token.token === context.req.headers['auth-token'] ||
     role.role === "admin"
   ) {
     const review = await getReviewById(id, context);
     const sql = `DELETE FROM review WHERE review.id = ?`;
     return new Promise((resolve, reject) => {
-      context.db.run(sql,[review.id], (err, rows) => {
+      context.db.run(sql,[review.id], (err:any) => {
         if (err) {
           reject(err);
         }
@@ -108,13 +115,3 @@ async function deleteReview(id, context) {
   }
   throw new Error("Unauthorized!");
 }
-
-module.exports = {
-  getReviews,
-  getReviewById,
-  getReviewsOfUserForMovie,
-  getReviewsOfMovie,
-  deleteReview,
-  updateReview,
-  createReview,
-};
